Updated a lot of stuff, Descrip has details
So I've moved all of the text image generators inside of the new data_grabber directory as well as combined the ffmpeg image with the selenium image and placed it there. This was necessary because the text image generators need ffmpeg to function and we need the file conversion process to happen when we collect the data.